% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeligData-package.R
\docType{package}
\name{ZeligData-package}
\alias{ZeligData}
\alias{ZeligData-package}
\title{ZeligData: Zelig Data Sets}
\description{
Datasets used in the Zelig package.
}
\author{
\strong{Maintainer}: Jeffrey Arnold \email{jeffrey.arnold@gmail.com}

Authors:
\itemize{
  \item Christine Choirat
  \item Christopher Gandrud
  \item James Honaker
  \item Kosuke Imai
  \item Gary King
  \item Olivia Lau
}

Other contributors:
\itemize{
  \item IQSS Harvard University [copyright holder]
}

}
\keyword{internal}
